chess_system/requirements.txt
python-chess>=1.999
pillow>=10.0
cairosvg>=2.7
